+++ src/pages/products/list/index.js
@@ -1,39 +1,12 @@
+import BasicPage from '../../basic-page';
+
 /**
  * Product list page
  */
-export default class Page {
-  subElements = {};
-  components = {};
+export default class extends BasicPage {
 
   async render() {
-    const element = document.createElement('div');
-
-    element.innerHTML = `
-      <div>
-        <h1>List page</h1>
-      </div>`;
-
-    this.element = element.firstElementChild;
-
-    this.initComponents();
-    await this.renderComponents();
-
-    return this.element;
-  }
-
-  initComponents() {
-  }
-
-  async renderComponents() {
-    for (const component of Object.values(this.components)) {
-      await component.render();
-      this.element.append(component.element);
-    }
-  }
-
-  destroy() {
-    for (const component of Object.values(this.components)) {
-      component.destroy();
-    }
+    this.element.innerHTML = `<h1>List page</h1>`;
+    return super.render();
   }
 }
