+++ src/pages/sales/index.js
@@ -1,8 +1,12 @@
+import BasicPage from '../basic-page';
+
 /**
  * Sales page
  */
-export default class Page {
-  async render() {
+export default class extends BasicPage {
 
+  async render() {
+    this.element.innerHTML = `<h1>Sales page</h1>`;
+    return super.render();
   }
 }
